+++ src/plugins/credential_management/server/crypto/crypto_cli.ts
@@ -9,5 +9,4 @@
  * GitHub history for details.
  */
 
-export { CryptographySingleton } from './singleton/cryptography_singleton';
-export { generateCryptoMaterials } from './crypto_cli';
+export { generateCryptoMaterials } from './singleton';
